Split the class objects into separate modules. Handled threading deadlocks.
